% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacyReports-package.R
\docType{package}
\name{pharmacyReporting-package}
\alias{pharmacyReporting}
\alias{pharmacyReporting-package}
\title{pharmacyReporting: Generate Reports For Pharmacy Inventory Project}
\description{
Using forecasting and stock control mechanisms.
}
\author{
\strong{Maintainer}: Chris Beeley \email{chris.beeley@gmail.com}

}
\keyword{internal}
