% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacyReports-package.R
\docType{package}
\name{pharmacyReports-package}
\alias{pharmacyReports}
\alias{pharmacyReports-package}
\title{pharmacyReports: Generate Reports For Pharmacy Inventory Project}
\description{
Using forecasting and stock control mechanisms.
}
\author{
\strong{Maintainer}: Chris Beeley \email{chris.beeley@gmail.com}

}
\keyword{internal}
